requirements.txt
streamlit
pandas
requests
google-generativeai
xlsxwriter
openpyxl